requirements.txt
Django==4.2
djangorestframework==3.15.2
mysqlclient==2.1
django-mysql==3.9
PyJWT==2.9.0